refact: closet router 분리 및 타입 정리
