+++ server/src/routes/comment.routes.ts
@@ -7,29 +7,29 @@ export const router = Router();
 router.put("/update/:shortId", async (req, res, next) => {
   const { content } = req.body;
   const { shortId } = req.params;
-  const tokenInfo = req.tokenInfo;
+  const email = req.email;
   let comment = {};
 
   try {
     comment = await Upment.findOne({ shortId });
     if (comment) {
       //작성자 검증
-      if (!tokenInfo) {
+      if (!email) {
         return next(new Error("로그인을 해주세요."));
       }
       const postUserId = await Upment.findOne({ shortId }).populate("author");
-      if (tokenInfo.email !== postUserId.author.email) {
+      if (email.email !== postUserId.author.email) {
         return next(new Error("작성자가 아닙니다!"));
       }
 
       await Upment.updateOne({ shortId }, { comment: content });
       res.status(200).json({ result: "수정이 완료되었습니다." });
     } else {
-      if (!tokenInfo) {
+      if (!email) {
         return next(new Error("로그인을 해주세요."));
       }
       const postUserId = await Downment.findOne({ shortId }).populate("author");
-      if (tokenInfo.email !== postUserId.author.email) {
+      if (email.email !== postUserId.author.email) {
         return next(new Error("작성자가 아닙니다!"));
       }
       await Downment.updateOne({ shortId }, { comment: content });
@@ -42,29 +42,29 @@ router.put("/update/:shortId", async (req, res, next) => {
 
 router.delete("/delete/:shortId", async (req, res, next) => {
   const { shortId } = req.params;
-  const tokenInfo = req.tokenInfo;
+  const email = req.email;
   let comment = {};
 
   try {
     comment = await Upment.findOne({ shortId });
     if (comment) {
       //작성자 검증
-      if (!tokenInfo) {
+      if (!email) {
         return next(new Error("로그인을 해주세요."));
       }
       const postUserId = await Upment.findOne({ shortId }).populate("author");
-      if (tokenInfo.email !== postUserId.author.email) {
+      if (email.email !== postUserId.author.email) {
         return next(new Error("작성자가 아닙니다!"));
       }
 
       await Upment.updateOne({ shortId }, { show: false });
       res.status(200).json({ result: "삭제가 완료되었습니다." });
     } else {
-      if (!tokenInfo) {
+      if (!email) {
         return next(new Error("로그인을 해주세요."));
       }
       const postUserId = await Downment.findOne({ shortId }).populate("author");
-      if (tokenInfo.email !== postUserId.author.email) {
+      if (email.email !== postUserId.author.email) {
         return next(new Error("작성자가 아닙니다!"));
       }
