+++ server/src/routes/index.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import * as users from "./user.routes";
-// import * as closet from "./closet.routes.js";
+import * as closet from "./closet.routes";
 // import * as posts from "./posts.routes.js";
 // import * as market from "./market.routes.js";
 // import * as like from "./like.routes.js";
@@ -8,14 +8,14 @@ import * as users from "./user.routes";
 // import * as search from "./search.routes.js";
 // import * as comment from "./comment.routes.js";
 
-import authmiddleware from "../util/authmiddleware.js";
+import authmiddleware from "../util/authmiddleware";
 
 console.log("routes loaded!!");
 
 const router = Router();
 
 router.use(users.path, users.router);
-// router.use(closet.path, authmiddleware, closet.router);
+router.use(closet.path, authmiddleware, closet.router);
 // router.use(posts.path, posts.router);
 // router.use(market.path, market.router);
 // router.use(like.path, authmiddleware, like.router);
