+++ server/src/routes/user.routes.ts
@@ -73,7 +73,7 @@ router.post("/login", login);
 // router.get("/mypost", authmiddleware, async (req, res, next) => {
 //   try {
 //     const postType = req.query.postType;
-//     const email = req.tokenInfo.email;
+//     const email = req.email.email;
 //     const page = Number(req.query.page);
 //     const perPage = Number(req.query.perPage);
 
@@ -99,7 +99,7 @@ router.post("/login", login);
 // // 내가 좋아요 누른 게시글 불러오기
 // router.get("/mylike", authmiddleware, async (req, res, next) => {
 //   try {
-//     const email = req.tokenInfo.email;
+//     const email = req.email.email;
 //     const page = Number(req.query.page);
 //     const perPage = Number(req.query.perPage);
 
@@ -127,7 +127,7 @@ router.post("/login", login);
 
 // router.put("/password/change", authmiddleware, async (req, res, next) => {
 //   try {
-//     const email = req.tokenInfo.email;
+//     const email = req.email.email;
 //     const { password } = req.body;
 
 //     let hashPassword = passwordHash(password);
