+++ server/src/types/express.d.ts
@@ -1,0 +1,9 @@
+import { Request } from "express";
+
+declare global {
+  namespace Express {
+    interface Request {
+      email: string;
+    }
+  }
+}
